chore: renamed feature
