+++ src/renderer/features/init-operation/ui/MultishardStaking/MultishardStaking.tsx
