+++ src/renderer/features/init-operation/ui/MultisigOperation/MultisigOperation.tsx
