+++ src/renderer/features/init-operation/ui/OperationFooter/OperationFooter.tsx
