+++ src/renderer/pages/Staking/Operations/StakeMore/InitOperation/InitOperation.tsx
@@ -16,7 +16,7 @@ import {
   validateBalanceForFeeDeposit,
 } from '../../common/utils';
 import { getSignatoryOption } from '@renderer/pages/Transfer/common/utils';
-import { OperationFooter, OperationHeader } from '@renderer/features/InitOperation';
+import { OperationFooter, OperationHeader } from '@renderer/features/init-operation';
 
 export type StakeMoreResult = {
   accounts: Account[];
